+++ src/main/java/searchengine/tools/indexing/ScrapingAction.java
@@ -94,7 +94,8 @@ public class ScrapingAction extends RecursiveAction {
 
 			try {
 				if (url.matches(URL_IS_VALID)
-						&& href.startsWith(siteUrl)
+						//Temp check for this site
+						&& (href.startsWith(siteUrl) | href.contains("https://my.tretyakov.ru/app/gallery") | href.contains("https://my.tretyakov.ru/app/masterpiece/"))
 						&& !href.contains("#")
 						&& !href.equals(url)
 						&& !newChildLinks.contains(href)